Massive overhaul to id record handling.
We no longer have separate fields for 'negative' values. Instead, if we start with an added value the record containing only that value will be made. Further additions will be made, as will further negations, in that same order. This prevents the previous situation of negative values always taking precendence over positive ones. If we start with a negative value, all values except for the negated one are included, and the same process happens with successive ones. We also removed counters to show specifically the number of each value type, as it's less relevant to this new design, and a pain to program.