Make the soft-necessitator options not affect local query results.
Some fairly complex changes were needed to fully implement this change. Primarily, it was necessary to avoid registering any empty or pure whitespace values, since they would nullify the results.

This is in effect, a welcome change, since it solves the problem, and makes the query field behavior more consistent. Otherwise, we could just treat it like a negated option.